+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 - Add support for MultiKey's in transaction simulations
 - Adds default implementation for `toString` and `toStringWithoutPrefix` for `Serializable`
 - Bump `@aptos-labs/aptos-cli` to `1.0.1`
+- [`Breaking`] Updated `AccountAddress.fromString` and `AccountAddress.from` to only accept SHORT strings that are 60-64 characters long by default (with the exception of special addresses). This can be adjusted using `maxMissingChars` which is set to `4` by default. If you would like to keep the previous behavior, set `maxMissingChars` to `63` for relaxed parsing.
 
 # 1.29.1 (2024-10-09)